initialize repository
Created the repository with individual scripts to insert the left half of a human, a t-shirt, a pair of jeans, and a full rigify metarig for the character. All positioned correctly and ready for a mirror modifier on the meshes, materials/shaders, and parenting all meshes to the armature in order to create animations.
These scripts are the general starting point ready for artistic customization to more quickly get a character up and running.